requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
async-timeout==5.0.1
attrs==25.3.0
discord.py==2.6.0
freeguy==4.0.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
python-dotenv==1.1.1
typing_extensions==4.14.1
yarl==1.20.1
